Switch some enums over to enum class to keep VS happy

+++ Engine/Graphics/Device/_vulkan/RenderPass.cpp
@@ -122,7 +122,7 @@ void RenderPass::Init(GFXDevice* pDevice, const RenderPassInitData &initData, ui
 			attachmentDescriptions[i].format = pDevice->GetPlatform().GetColorFormat(in.format.eColor);
 			break;
 		case RenderPassDecl::ATTACH_DEPTH:
-			attachmentDescriptions[i].format = gDepthFormatViewMap[in.format.eDepth];
+			attachmentDescriptions[i].format = gDepthFormatViewMap[(uint32)in.format.eDepth];
 			break;
 		default:
 			ASSERT(false);
